+++ src/apps/convex/arbitrum/convex.booster.contract-position-fetcher.ts
@@ -3,7 +3,7 @@ import { PositionTemplate } from '~app-toolkit/decorators/position-template.deco
 import { ConvexBoosterContractPositionFetcher } from '../common/convex.booster.contract-position-fetcher';
 
 @PositionTemplate()
-export class EthereumConvexBoosterContractPositionFetcher extends ConvexBoosterContractPositionFetcher {
+export class ArbitrumConvexBoosterContractPositionFetcher extends ConvexBoosterContractPositionFetcher {
   groupLabel = 'Booster';
   boosterAddress = '0xf403c135812408bfbe8713b5a23a04b3d48aae31';