+++ src/apps/good-ghosting/common/good-ghosting.game.games.api-source.ts
@@ -11,7 +11,7 @@ import { NetworkId, getGameVersionType, RewardType, BASE_API_URL, GamesResponse 
 @Injectable()
 export class GoodGhostingGameGamesApiSource {
   @CacheOnInterval({
-    key: `studio:${GOOD_GHOSTING_DEFINITION.id}:${GOOD_GHOSTING_DEFINITION.groups.game}:addresses`,
+    key: `studio:${GOOD_GHOSTING_DEFINITION.id}:${GOOD_GHOSTING_DEFINITION.groups.game.id}:addresses`,
     timeout: 15 * 60 * 1000,
     failOnMissingData: false,
   })