+++ src/apps/maple/helpers/maple.cache-manager.ts
@@ -9,8 +9,8 @@ import { MAPLE_DEFINITION } from '../maple.definition';
 type MapleAllPoolsResponse = {
   allPools?: {
     list?: {
+      id: string;
       poolName: string;
-      contractAddress: string;
       symbol: string;
       lpApy: string;
       farmingApy: string;
@@ -28,8 +28,8 @@ const ALL_POOLS_QUERY = gql`
   {
     allPools {
       list {
+        id
         poolName
-        contractAddress
         symbol
         lpApy
         farmingApy
@@ -60,7 +60,7 @@ export class MapleCacheManager {
     return (pairsData.allPools?.list ?? []).map(v => ({
       poolName: v.poolName,
       apy: Number(v.lpApy) + Number(v.farmingApy),
-      poolAddress: v.contractAddress.toLowerCase(),
+      poolAddress: v.id.toLowerCase(),
       farmAddress: v.stakeRewards.id.toLowerCase(),
     }));
   }