+++ src/apps/stake-dao/contracts/ethers/factories/index.ts
@@ -2,4 +2,6 @@
 /* tslint:disable */
 /* eslint-disable */
 export { StakeDaoCurvePool__factory } from './StakeDaoCurvePool__factory';
+export { StakeDaoFarm__factory } from './StakeDaoFarm__factory';
 export { StakeDaoGauge__factory } from './StakeDaoGauge__factory';
+export { StakeDaoVotingEscrow__factory } from './StakeDaoVotingEscrow__factory';